Introduce `sm_pt_resupply` and `sm_pt_allow_instant_resupply`
A "resupply" bind consists of loading an item preset (by doing
`load_itempreset 0`), and having `tf_respawn_on_loadoutchanges` set to
`1`. This, however, results in a delayed resupply and respawn.

The delay is due to the how `tf_respawn_on_loadoutchanges` is
implemented: the client will fire a `k_EMsgGCRespawnPostLoadoutChange`
message to the GC, which will reach the server after some indirection,
which delays the result.

We slightly re-implement the logic from the game, allowing for
near-instantaneous resupply and respawn, without worrying about loadout
changes.