Update pt_cup.cfg
Removed redudant commands as RGL's base config has been updated to include most of the commands we have in our config. Also changed winlimit from 2 to 3 as it was inconsistent with the message at the bottom saying it was 3. Don't know if this was intentional or not, can be changed back if so