CFG modifications (#27)
* Heavily modify cfgs to include tf passtime defaults

* Update pt_cup.cfg (buff overtime idle time)

* Update pt_global_official.cfg (buff overtime idle time)

* Update pt_global_pug.cfg

* remove redundant cfg values

* Update pt_global_pug.cfg

Removed redudant commands as RGL's base config has been updated to include most of the commands we have in our config

* Update pt_global_pug.cfg

Readded stuff I didn't mean to remove

* Update pt_global_official.cfg

Removed redudant commands as RGL's base config has been updated to include most of the commands we have in our config + fixed typo

* Update pt_global_pug.cfg

Removed extra comment because I can't read

* Update pt_global_official.cfg

* Update pt_cup.cfg

Removed redudant commands as RGL's base config has been updated to include most of the commands we have in our config. Also changed winlimit from 2 to 3 as it was inconsistent with the message at the bottom saying it was 3. Don't know if this was intentional or not, can be changed back if so

---------

Co-authored-by: Speakers <92229336+SpeakersVolumeUp@users.noreply.github.com>